docs: update batch-usage reflect the correct usage (#93)
The documentation incorrectly stated that `database.batch()` would return a `Batch` object. The recommended way to get a `Batch` object is to use the `BatchCheckout` as a context manager in a `with` statement. The documentation has been update to reflect this.

Documentation links and minor grammar fixes are also included.