test: fix credential scope assertions (#223)
The assertions for credential scope in the `client` unit tests were broken by [a PR in the auth library](https://github.com/googleapis/google-auth-library-python/pull/665). This does raise the question of whether we should be asserting the scopes like this in this library.

This PR fixes the assertions. Removal of these assertions can be done in a separate PR if it is decided they don't belong in this library.