refactor(dbapi): Remove most special handling of INSERTs
For historical reasons it seems the INSERT codepath and that for
UPDATE/DELETE were separated, but today there appears to be no practical
differences in how these DML statements are handled. This change removes
most of the special handling for INSERTs and uses existing methods for
UPDATEs/DELETEs instead. The one remaining exception is the automatic
addition of a WHERE clause to UPDATE and DELETE statements lacking one,
which does not apply to INSERT statements.