feat(default_retention_period): Changes to copy backup path to support null expire time
With the default retention period feature, null value for expiration time
is a valid value. If the expiration time of the destination backup is not
specified (null) at the time of copy backup then the destination backup
will be created with the expiration time of the source backup.

In this commit, the precondition check on expiration time in the copy
backup path is removed.