perf: add option for last_statement
Adds an option to indicate that a statement is the last statement in a read/write
transaction. Setting this option allows Spanner to optimize the execution of the
statement, and defer some validations until the Commit RPC that should follow
directly after this statement.

The last_statement option is automatically used by the dbapi driver when a
statement is executed in autocommit mode.