feat: support configuring logger in dbapi kwargs
Allow the kwargs for dbapi connections to inlcude a logger, and use this as the
logger for the database that is used.

Also set a default logger that only logs at WARN level for the mock server tests
to stop them from spamming the test log with a bunch of "Created multiplexed session."
messages that are logged at INFO level.